models/HFALaMa/requirements.txt
torch==2.4.1
torchvision==0.19.1
numpy==1.24.4
opencv-python==4.10.0.84
Pillow==9.4.0
einops==0.4.1
hydra-core==1.3.2
omegaconf==2.3.0
PyYAML==6.0.2
tqdm==4.67.1
matplotlib==3.7.5